[maui] Work on authentication flow.
Separated responsibility between ShellPage and Login page. ShellPage should be able to transiently reuthenitcate a user after token expiration and automatically sign user in without having to go to Login page.

while the Login page provides a distinct separate UX when user logs in for the first time or logs out
