+++ src/MvpCompanion.Maui/ShellPage.xaml.cs
@@ -1,14 +1,13 @@
 ﻿using MvpApi.Common.CustomEventArgs;
-using MvpApi.Common.Interfaces;
-using MvpApi.Common.Models.Navigation;
 using MvpApi.Services.Apis;
 using MvpApi.Services.Utilities;
 using MvpCompanion.Maui.ViewModels;
+using MvpCompanion.Maui.Views;
 using System.Text.Json;
 
 namespace MvpCompanion.Maui;
 
-public partial class ShellPage : Shell, INavigationHandler
+public partial class ShellPage : Shell
 {
     //private readonly WebView _webView;
     private readonly ShellViewModel _viewModel;
@@ -17,78 +16,250 @@ public partial class ShellPage : Shell, INavigationHandler
 	{
 		InitializeComponent();
         _viewModel = new ShellViewModel();
-        this.BindingContext = _viewModel;
+        BindingContext = _viewModel;
 
         if (Device.Idiom == TargetIdiom.Phone)
         {
             CurrentItem = PhoneTabs;
         }
 
-        Routing.RegisterRoute("login", typeof(Views.Login));
-        Routing.RegisterRoute("home", typeof(Views.Home));
-        Routing.RegisterRoute("upload", typeof(Views.Upload));
-        Routing.RegisterRoute("detail", typeof(Views.Detail));
-        Routing.RegisterRoute("help", typeof(Views.Help));
-        Routing.RegisterRoute("profile", typeof(Views.Profile));
-        Routing.RegisterRoute("about", typeof(Views.About));
-        Routing.RegisterRoute("settings", typeof(Views.Settings));
-        
-        _viewModel = new ShellViewModel
-        {
-            NavigationHandler = this
-        };
-
-        this.GoToAsync("login");
-
-        //_webView = new WebView();
-        //_webView.Navigated += WebView_OnNavigated;
+        Routing.RegisterRoute("login", typeof(Login));
+        //Routing.RegisterRoute("home", typeof(Views.Home));
+        Routing.RegisterRoute("home/detail", typeof(Detail));
+        Routing.RegisterRoute("upload", typeof(Upload));
+        //Routing.RegisterRoute("profile", typeof(Profile));
+        Routing.RegisterRoute("help", typeof(Help));
+        Routing.RegisterRoute("settings", typeof(Settings));
+        //Routing.RegisterRoute("about", typeof(About));
     }
 
-	public void LoadView(ViewType viewType)
+    protected override async void OnAppearing()
     {
-        switch (viewType)
+        base.OnAppearing();
+
+        var userAuthRefreshed = await SignInAsync();
+
+        if (userAuthRefreshed)
         {
-            case ViewType.Home:
-                GoToAsync("///home");
-                break;
-            case ViewType.Upload:
-                GoToAsync("///upload");
-                break;
-            case ViewType.Detail:
-                GoToAsync("///detail");
-                break;
-            case ViewType.Profile:
-                GoToAsync("///profile");
-                break;
-            case ViewType.Help:
-                GoToAsync("///help");
-                break;
-            case ViewType.Settings:
-                GoToAsync("///settings");
-                break;
-            case ViewType.About:
-                GoToAsync("///about");
-                break;
-            default:
-                throw new ArgumentOutOfRangeException(nameof(viewType), viewType, null);
+            await GoToAsync("home");
+        }
+        else
+        {
+            await GoToAsync("login?operation=signin");
         }
     }
-
-    void TapGestureRecognizer_Tapped(System.Object sender, System.EventArgs e)
+    
+    private void TapGestureRecognizer_Tapped(Object sender, EventArgs e)
     {
         GoToAsync("///settings");
     }
+    
+    #region Authentication
 
-    protected override bool OnBackButtonPressed()
+    
+
+    public async Task<bool> SignInAsync()
     {
-        // TODO Replace SideDrawer
-        //if (SideDrawer.MainContent.GetType() != typeof(HomeView))
-        //{
-        //	SideDrawer.MainContent = new HomeView();
-        //	return true;
-        //}
+        try
+        {
+            var refreshToken = StorageHelpers.Instance.LoadToken("refresh_token");
 
-        return base.OnBackButtonPressed();
-	}
+            // If refresh token is available, the user has previously been logged in and we can get a refreshed access token immediately
+            if (!string.IsNullOrEmpty(refreshToken))
+            {
+                _viewModel.IsBusy = true;
+                _viewModel.IsBusyMessage = "refreshing session...";
+
+                var authorizationHeader = await RequestAuthorizationAsync(refreshToken, true);
+
+                if (!string.IsNullOrEmpty(authorizationHeader))
+                {
+                    await InitializeMvpApiAsync(authorizationHeader);
+                    return true;
+                }
+            }
+
+            return false;
+        }
+        catch (Exception e)
+        {
+            await e.LogExceptionAsync();
+            throw;
+        }
+        finally
+        {
+            _viewModel.IsBusy = false;
+            _viewModel.IsBusyMessage = "";
+        }
+    }
+
+    public async Task SignOutAsync()
+    {
+        try
+        {
+            // Indicate to user we are signing out
+            _viewModel.IsBusy = true;
+            _viewModel.IsBusyMessage = "logging out...";
+
+            // Erase cached tokens
+            _viewModel.IsBusyMessage = "deleting cache files...";
+            StorageHelpers.Instance.DeleteToken("access_token");
+            StorageHelpers.Instance.DeleteToken("refresh_token");
+
+            // Delete profile photo file
+            if (File.Exists((Current.BindingContext as ShellViewModel).ProfileImagePath))
+            {
+                (Current.BindingContext as ShellViewModel).IsBusyMessage = "deleting profile photo file...";
+                File.Delete((Current.BindingContext as ShellViewModel).ProfileImagePath);
+            }
+
+            // Clean up profile objects
+            _viewModel.IsBusyMessage = "resetting profile...";
+            (Current.BindingContext as ShellViewModel).Mvp = null;
+            (Current.BindingContext as ShellViewModel).ProfileImagePath = "";
+        }
+        catch (Exception ex)
+        {
+            await ex.LogExceptionAsync();
+            await DisplayAlert("Sign out Error", ex.Message, "ok");
+        }
+        finally
+        {
+            // Hide busy indicator
+            _viewModel.IsBusy = false;
+            _viewModel.IsBusyMessage = "";
+
+            // Toggle flag
+            (Current.BindingContext as ShellViewModel).IsLoggedIn = false;
+
+            await GoToAsync("login?operation=signout");
+        }
+    }
+    
+    public async Task InitializeMvpApiAsync(string authorizationHeader)
+    {
+        _viewModel.IsBusy = true;
+        _viewModel.IsBusyMessage = "authenticating...";
+
+        // remove any previously wired up event handlers
+        if (App.ApiService != null)
+        {
+            App.ApiService.AccessTokenExpired -= ApiService_AccessTokenExpired;
+            App.ApiService.RequestErrorOccurred -= ApiService_RequestErrorOccurred;
+        }
+
+        App.ApiService = new MvpApiService(authorizationHeader);
+
+        App.ApiService.AccessTokenExpired += ApiService_AccessTokenExpired;
+        App.ApiService.RequestErrorOccurred += ApiService_RequestErrorOccurred;
+
+        (Current.BindingContext as ShellViewModel).IsLoggedIn = true;
+
+        _viewModel.IsBusyMessage = "downloading profile info...";
+        (Current.BindingContext as ShellViewModel).Mvp = await App.ApiService.GetProfileAsync();
+
+        _viewModel.IsBusyMessage = "downloading profile image...";
+        (Current.BindingContext as ShellViewModel).ProfileImagePath = await App.ApiService.DownloadAndSaveProfileImage();
+
+        // Navigation is now performed where ever the initialization request was called from.
+        // This allows for individually determining if navigation is needed
+        // await Shell.Current.GoToAsync("home");
+
+        _viewModel.IsBusyMessage = "";
+        _viewModel.IsBusy = false;
+    }
+
+    // WebView logic
+
+
+    private readonly string _redirectUrl = "https://login.live.com/oauth20_desktop.srf";
+    private readonly string _accessTokenUrl = "https://login.live.com/oauth20_token.srf";
+    private static readonly string _clientId = "090fa1d9-3d6f-4f6f-a733-a8b8a3fe16ff";
+
+    public async Task<string> RequestAuthorizationAsync(string authCode, bool isRefresh = false)
+    {
+        try
+        {
+            using var client = new HttpClient();
+
+            // Construct the Form content, this is where I add the OAuth token (could be access token or refresh token)
+            var postContent = new FormUrlEncodedContent(new List<KeyValuePair<string, string>>
+            {
+                new("client_id", _clientId),
+                new("grant_type", isRefresh ? "refresh_token" : "authorization_code"),
+                new(isRefresh ? "refresh_token" : "code", authCode.Split('&')[0]),
+                new("redirect_uri", _redirectUrl)
+            });
+
+            // Variable to hold the response data
+            var responseTxt = "";
+
+            // post the Form data
+            using (var response = await client.PostAsync(new Uri(_accessTokenUrl), postContent))
+            {
+                // Read the response
+                responseTxt = await response.Content.ReadAsStringAsync();
+            }
+
+            // Deserialize the parameters from the response
+            var tokenData = JsonSerializer.Deserialize<Dictionary<string, string>>(responseTxt);
+
+            if (tokenData.ContainsKey("access_token"))
+            {
+                StorageHelpers.Instance.StoreToken("access_token", tokenData["access_token"]);
+                StorageHelpers.Instance.StoreToken("refresh_token", tokenData["refresh_token"]);
+
+                // We need to prefix the access token with the token type for the auth header. 
+                // Currently this is always "bearer", doing this to be more future proof
+                var tokenType = tokenData["token_type"];
+                var cleanedAccessToken = tokenData["access_token"].Split('&')[0];
+
+                // set public property that is "returned"
+                return $"{tokenType} {cleanedAccessToken}";
+            }
+        }
+        catch (HttpRequestException e)
+        {
+            await e.LogExceptionAsync();
+            await DisplayAlert("Error", $"Something went wrong signing you in, try again. {e.Message}", "ok");
+        }
+        catch (Exception e)
+        {
+            await e.LogExceptionAsync();
+            await DisplayAlert("Error", $"Something went wrong signing you in, try again. {e.Message}", "ok");
+        }
+
+        return null;
+    }
+
+    private async void ApiService_AccessTokenExpired(object sender, ApiServiceEventArgs e)
+    {
+        if (e.IsTokenRefreshNeeded)
+        {
+            var userAuthRefreshed = await SignInAsync();
+
+            if (!userAuthRefreshed)
+            {
+                await GoToAsync("login?signin");
+            }
+        }
+    }
+
+    private async void ApiService_RequestErrorOccurred(object sender, ApiServiceEventArgs e)
+    {
+        var message = "Unknown Server Error";
+
+        if (e.IsBadRequest)
+        {
+            message = e.Message;
+        }
+        else if (e.IsServerError)
+        {
+            message = e.Message + "\r\n\nIf this continues to happen, please open a GitHub Issue and we'll investigate further (find the GitHub link on the About page).";
+        }
+
+        await DisplayAlert("MVP API Request Error", message, "ok");
+    }
+
+    #endregion
 }
-
