+++ src/MvpCompanion.Maui/ViewModels/LoginViewModel.cs
@@ -1,12 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using CommonHelpers.Common;
+﻿using CommonHelpers.Common;
 
 namespace MvpCompanion.Maui.ViewModels;
 
 public class LoginViewModel : ViewModelBase
 {
+    public LoginViewModel()
+    {
+    }
 }
